JuniorTask/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-filter==2.2.0
django-rest==0.8.7
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
psycopg2-binary==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
zipp==3.8.0
